requirements.txt
Django==3.0.2
psycopg2==2.8.4
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
faker==4.0.0